refactor: split report failure digest helpers
